requirements.txt
discord
googlesearch-python
gspread
oauth2client
pytz
IMDbPY
bs4
requests